requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymongo[srv]
pydantic
python-multipart
typing-extensions
python-jose[cryptography]
passlib[bcrypt]